feat: implementando permissões

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\Models\Role;
 use App\Models\User;
 use App\Models\UserProfile;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -47,6 +48,7 @@ class UserController extends Controller
 
     public function edit(User $user)
     {
+        Gate::authorize('edit', User::class);
         $user->load('profile', 'interests');
         $roles = Role::all();
         return view('users.edit', compact('user', 'roles'));
@@ -54,6 +56,7 @@ class UserController extends Controller
 
     public function update(User $user, Request $request)
     {
+        Gate::authorize('edit', User::class);
         $input = $request->validate([
             'name' => 'required',
             'email' => 'required|email',
@@ -67,6 +70,7 @@ class UserController extends Controller
 
     public function updateProfile(User $user, Request $request)
     {
+        Gate::authorize('edit', User::class);
         $input = $request->validate([
             'type' => 'required',
             'address' => 'nullable',
@@ -81,6 +85,7 @@ class UserController extends Controller
 
     public function updateInterests(User $user, Request $request)
     {
+        Gate::authorize('edit', User::class);
         $input = $request->validate([
             'interests' => 'nullable|array',
         ]);
@@ -95,6 +100,7 @@ class UserController extends Controller
 
     public function updateRoles(User $user, Request $request)
     {
+        Gate::authorize('edit', User::class);
 
         $input = $request->validate([
             'roles' => 'required|array',
@@ -106,6 +112,7 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        Gate::authorize('destroy', User::class);
         $user->delete();
         return redirect()->route('users.index')->with('status', 'Usuário deletado com sucesso!');
     }
